c12e-ci: prep /root as a data container for dev builds
We're already doing this for the nix-data container.  Also, this protects us
from Docker runs making root-owned files in bind-mounted directories.
Actually, this still happens in the workspace, which is why the cleanup stage
does a chown.  Maybe there's a better way to manage ownership of files in the
root directory.