c12e-ci: revert tagging behavior to prior semantics
In an effort to simplify the script, I changed how it tagged images for "dev
mode" only to want the old semantics back after thinking about it.

The problem is that for any built artifact (in this case a Docker image), we
only want one official build of it.  In the case of Nix, it wouldn't matter how
many different times we built an artifact, because we'd trust in the
determinism of a Nix build.  But Docker doesn't give us that guarantee.

So if I come in as a developer and publish a dev version of a tag.  I want that
in a different namespace, so the official GoCD build won't fail if it comes
alone and builds the exact same commit.