c12e-ci: check if already published works again!
I finally put in the necessary curl and jq calls to authenticate correctly
with the Docker registry so that the check of whether an image is published or
not actually works.

Johan has shortcutted out the check with a "return 0", ostensibly with the
idea that this work could be finished later.