c12e-ci: made nix-data a mandatory volume for all builds
Unfortunately, volumes are not cummulative across multiple compose files.  So
the "nix-data" requirement can not be refactored out from the whole
"volumes_from" list.

That's the bad part -- that Compose files don't compose as well as one would
like.  The good part is that this shouldn't be too much of a liability at
runtime.