c12e-ci: not leave dev-home mounted to $HOME
We're using Docker data containers, but they're kind of encoded into the
normal Docker command set.  To make a proper data container, you first need to
create it with the volumes you mean to expose, and then attach to it with
another container to populate it.  I was merging these two steps into one,
which "works" but leaves you entire home directory mounted in the data
container's /mnt directory, which is needless exposure.