c12e-ci: move to "Dockerfile.c12e-ci.*" files
It used to be that the script would build all the found Dockerfiles, but
that left the question of what the name of the built image should be.

To facilitate this, we were (previous to this commit) specifying the
image name with a "-n" option.

Now, we're going to encode the image name in the Dockerfile filename.
Only files of name "Dockerfile.c12e-ci.*" will be built, where the repo
name will be pulled extracted as "$REPO" from
"Dockerfile.c12e-ci.$REPO".

This will enable GoCD to use a more consistent template across more
builds, because the "-n" switch won't need to be parameterized.