c12e-ci: remove redundant "dev-latest" tag
We already have the "latest-dev" tag.  It's very confusing to have two
tags with the same semantics.  We must topple ignorance, and not live in
the fear that our legacy imposes.  Let's keep things clean.

A follow-up to this activity would be to go through all the repositories
and purge all the "dev-latest" tags.